+++ spring-hw-11/README.md
@@ -2,6 +2,3 @@
 
 ### Примечания/вопросы:
 
-- Убрал локализацию, так как понял, что на её реализацию уйдёт больше времени чем на сам реактивный стек
-- Пришлось создать GlobalErrorWebExceptionHandler для глобального перехвата ошибок контроллера. Не знаю в каком месте такие штуки размещать, оставил в конфиге
-- После предыдущего целенаправленного разнесение логики по слоям, меня по-прежнему коробит идея засовывания всего кода в контроллер. Я избавился от сервисов как было рекомендовано, но код всё равно вынес по индивидуальным методам вместо того, что бы пихать всю в основном бин, иначе там получается совсем уж чудовище
